cloudwatch-agent/requirements.txt
aws-cdk.aws-autoscaling
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-logs-destinations
aws-cdk.aws-s3-assets
aws-cdk.core
